+++ scripts/start.js
@@ -146,7 +146,7 @@ function runDevServer(port) {
     historyApiFallback: true,
     hot: true, // Note: only CSS is currently hot reloaded
     publicPath: config.output.publicPath,
-    quiet: true
+    quiet: true,
   }).listen(port, (err, result) => {
     if (err) {
       return console.log(err);