+++ src/features/Settings/components/ChangePassword/index.js
@@ -2,11 +2,10 @@ import { Component, PropTypes } from 'react';
 
 import styles from './styles.less';
 
+import CardWithTitle from 'common/layouts/CardWithTitle';
 import Textbox from 'common/components/Textbox';
-import Card from 'common/components/Card';
 import Button from 'common/components/Button';
 import PasswordForm from 'common/components/PasswordForm';
-import Message from 'common/components/Message';
 
 export default class ChangePassword extends Component {
   static propTypes = {
@@ -41,42 +40,40 @@ export default class ChangePassword extends Component {
   };
 
   render () {
+    const { message, error, busy, valid } = this.props;
+    const { currentPassword, password, passwordConfirm } = this.state;
+
     return (
-      <span>
-        <h2>Change password</h2>
-        <Card size="medium">
-          {this.props.message && <Message type="success">{this.props.message}</Message>}
+      <CardWithTitle size="medium" title="Change password" message={message}>
+        <form onSubmit={this.changePassword}>
+          <Textbox
+            required
+            id="currentPassword"
+            placeholder="Current password"
+            type="password"
+            value={currentPassword}
+            onChange={this.fieldChanged}
+          />
 
-          <form onSubmit={this.changePassword}>
-            <Textbox
-              required
-              id="currentPassword"
-              placeholder="Current password"
-              type="password"
-              value={this.state.currentPassword}
-              onChange={this.fieldChanged}
-            />
+          <PasswordForm
+            onFieldChange={this.fieldChanged}
+            valid={valid}
+            passwordValue={password}
+            passwordConfirmValue={passwordConfirm}
+            error={error}
+          />
 
-            <PasswordForm
-              onFieldChange={this.fieldChanged}
-              valid={this.props.valid}
-              passwordValue={this.state.password}
-              passwordConfirmValue={this.state.passwordConfirm}
-              error={this.props.error}
-            />
-
-            <div className={styles.buttons}>
-              <Button
-                busy={this.props.busy}
-                type="primary"
-                disabled={!this.props.valid}
-              >
-                CHANGE
-              </Button>
-            </div>
-          </form>
-        </Card>
-      </span>
+          <div className={styles.buttons}>
+            <Button
+              busy={busy}
+              type="primary"
+              disabled={!valid}
+            >
+              CHANGE
+            </Button>
+          </div>
+        </form>
+      </CardWithTitle>
     );
   }
 }