+++ src/lib/add-event.js
@@ -1,4 +1,0 @@
-export default function addEvent (event, cb) {
-  window.addEventListener(event, cb, false);
-  return () => window.removeEventListener(event, cb, false);
-}